+++ foxes/input/states/create/__init__.py
@@ -3,4 +3,5 @@ from .random_abl_states import write_random_abl_states as write_random_abl_state
 from .random_timeseries import random_timseries_data as random_timseries_data
 from .dataset_mean import create_dataset_mean as create_dataset_mean
 from .wrf2foxes import wrf2foxes as wrf2foxes
+from .era52foxes import era52foxes as era52foxes
 from .icon_dream import iconDream2foxes as iconDream2foxes